Did refactoring
