+++ Makefile
@@ -3,8 +3,8 @@ NAME = webserv
 OBJ_DIR = obj
 
 SOURCES = Webserv.cpp WebservInit.cpp WebservEvents.cpp WebservCgi.cpp WebservUtils.cpp \
-		  Logger.cpp Request.cpp Response.cpp \
-		  main.cpp 
+		  Response.cpp ResponseConsts.cpp ResponseDirectory.cpp ResponseUtils.cpp \
+		  Logger.cpp Request.cpp main.cpp 
 
 OBJECTS = $(addprefix $(OBJ_DIR)/, $(SOURCES:.cpp=.o))
