+++ Response.cpp
@@ -1,38 +1,5 @@
 #include "Response.hpp"
 
-
-const map_int_str Response::_response_codes = {
-	{200, "200 OK"},
-	{403, "403 Forbidden"},
-	{404, "404 Not Found"},
-	{405, "405 Method Not Allowed"},
-	{413, "413 Request Entity Too Large"},
-	{500, "500 Internal Server Error"}
-};
-
-const map_int_str Response::_error_pages = {
-	{403, "./default_pages/403.html"},
-	{404, "./default_pages/404.html"},
-	{405, "./default_pages/405.html"},
-	{413, "./default_pages/413.html"},
-	{500, "./default_pages/500.html"},
-	{0, "./default_pages/unknown.html"}
-};
-
-const map_str_str Response::_mime_types = {
-	{"html", "text/html"},
-	{"css", "text/css"},
-	{"js", "text/javascript"},
-	{"jpg", "image/jpeg"},
-	{"jpeg", "image/jpeg"},
-	{"png", "image/png"},
-	{"ico", "image/x-icon"},
-	{"json", "application/json"},
-	{"pdf", "application/pdf"},
-	{"zip", "application/zip"},
-	{"", "text/plain"}
-};
-
 Response::Response( void ) : logger(Logger::getInstance()) {
 	location = nullptr;
 }
@@ -66,19 +33,6 @@ int Response::prepareResponse( const std::string& file_path, size_t status_code 
 	}
 	_prepareStaticFile(full_path, extension, status_code);
 	return 1;
-}
-
-bool Response::_checkIfDirectory( const std::string& file_path, const std::string& full_path ) {
-	if (full_path.back() == '/' && _isDirectory(full_path)) {
-		if (location->autoindex == 1) {
-			_generateDirectoryList(full_path, file_path);
-		} else {
-			prepareResponseError(403);
-		}
-		return true;
-	} else {
-		return false;
-	}
 }
 
 int Response::_checkCgiAccess( const std::string& full_path ) {
@@ -145,14 +99,6 @@ int Response::_saveResponsePage( std::string& filepath, int status_code ) {
 	return 0;
 }
 
-std::string Response::_getFileExtension( const std::string& filepath ) {
-	size_t pos = filepath.rfind('.');
-	if (pos == std::string::npos) {
-		return "";
-	}
-	return filepath.substr(pos + 1);
-}
-
 std::string Response::_getHtmlHeader( size_t content_length, size_t status_code,
 									  const std::string& extension ) {
 	std::string header = "HTTP/1.1 ";
@@ -172,82 +118,6 @@ std::string Response::_getHtmlHeader( size_t content_length, size_t status_code,
 	return header;
 }
 
-//#include <sys/stat.h>
-int Response::_isDirectory( const std::string& full_path ) {
-	struct stat path_stat;
-	if (stat(full_path.c_str(), &path_stat) == 0) {
-		return (S_ISDIR(path_stat.st_mode));
-	}
-	return 0;
-}
-
-void Response::_generateDirectoryList( const std::string& dir_path,
-									   const std::string& file_path ) {
-	std::ostringstream html;
-	html << "<html><body><h1>Index of " << file_path << "</h1>\n<hr><pre><table>\n";
-	DIR *dir = opendir(dir_path.c_str());
-	if (dir == nullptr) {
-		prepareResponseError(403);
-		logger.warning("Failed to open directory" + dir_path);
-		return;
-	}
-	struct dirent *entry;
-	while ((entry = readdir(dir)) != nullptr) {
-		html << _getEntryLine(entry, dir_path);
-	}
-	closedir(dir);
-	html << "</table>\n</pre><hr></body>\n</html>\n";
-	full_response = "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nContent-Length: ";
-	full_response += std::to_string(html.str().size())+ "\r\n\r\n" + html.str();
-}
-
-// std::string Response::_getModTime( const std::string& path ) {
-// 	struct stat file_stat;
-// 	if (stat(path.c_str(), &file_stat) == 0) {
-// 		struct tm *tm = localtime(&file_stat.st_mtime);
-// 		char timebuf[80];
-// 		strftime(timebuf, sizeof(timebuf), "%d-%b-%Y %H:%M", tm);
-// 		return std::string(timebuf);
-// 	}
-// 	return "Unknown";
-// }
-
-std::string Response::_getEntryLine( struct dirent* entry, const std::string& dir_path ) {
-	std::string name, full_path, size, mod_time;
-	name = entry->d_name;
-	full_path = dir_path + name;
-	if (name == ".") {
-		return "";
-	}
-	if (name == "..") {
-		name += "/";
-	} else {
-		_getEntryStats(full_path, size, mod_time);
-		if (entry->d_type == DT_DIR) {
-			name += "/";
-			size = "-";
-		}
-	}
-	return std::string("<tr><td style=\"width:70%\"><a href=\"") + name + "\">" + 
-		   name + "</a></td>" + "<td style=\"width:20%\">" + mod_time + "</td>" +
-		   "<td align=\"right\">" + size + "</td></tr>\n";
-
-}
-
-void Response::_getEntryStats( const std::string& path, std::string& size, std::string& mod_time ) {
-	struct stat entry_stat;
-	if (stat(path.c_str(), &entry_stat) == -1) {
-		size = "Unknown";
-		mod_time = "Unknown";
-		return;
-	}
-	struct tm *tm = localtime(&entry_stat.st_mtime);
-	char timebuf[80];
-	strftime(timebuf, sizeof(timebuf), "%d-%b-%Y %H:%M", tm);
-	mod_time = std::string(timebuf);
-	size = std::to_string(entry_stat.st_size);
-}
-
 void Response::handleCgiResponse( void ) {
 	if (full_response.size() < 8 || full_response.compare(0, 8, "Status: ") != 0) {
 		full_response.insert(0, "HTTP/1.1 200 OK\r\n");
@@ -262,14 +132,5 @@ void Response::handleCgiResponse( void ) {
 		prepareResponseError(status_code);
 	} else {
 		full_response.replace(0, 7, "HTTP/1.1");
-	}
-}
-
-int Response::_stringToInt( const std::string& str ) {
-	try {
-		int status_code = std::stoi(str);
-		return status_code;
-	} catch (const std::invalid_argument&) {
-		return -1;
 	}
 }
