+++ Response.hpp
@@ -21,20 +21,24 @@ private:
 	static const map_int_str _error_pages;
 	static const map_str_str _mime_types;
 
-	bool _checkIfDirectory( const std::string& file_path, const std::string& full_path );
+	// Response.cpp
 	int _checkCgiAccess( const std::string& full_path );
 	void _prepareStaticFile(const std::string& path, const std::string& extension,
 							size_t status_code );
 	std::string _getErrorPagePath(const map_int_str& error_pages, size_t status_code);
 	int _saveResponsePage( std::string& filepath, int status_code );
-	std::string _getFileExtension( const std::string& filepath );
 	std::string _getHtmlHeader( size_t content_length, size_t status_code,
 								const std::string& extension );
+
+	// ResponseDirectory.cpp
+	bool _checkIfDirectory( const std::string& file_path, const std::string& full_path );
 	int _isDirectory( const std::string& full_path );
 	void _generateDirectoryList( const std::string& dir_path, const std::string& file_path );
-	// std::string _getModTime( const std::string& path );
 	std::string _getEntryLine( struct dirent* entry, const std::string& dir_path );
 	void _getEntryStats( const std::string& path, std::string& size, std::string& mod_time );
+
+	// ResponseUtils.cpp
+	std::string _getFileExtension( const std::string& filepath );
 	int _stringToInt( const std::string& str );
 
 
@@ -49,6 +53,7 @@ public:
 	Location*	location;
 	Logger&		logger;
 
+	// Response.cpp
 	int prepareResponse( const std::string& file_path, size_t status_code = 200 );
 	void prepareResponseError( size_t status_code );
 	void handleCgiResponse( void );
