+++ Webserv.cpp
@@ -410,174 +410,6 @@ void Webserv::_checkTimeouts( void ) {
 	}
 }
 
-// //#include <sys/stat.h>
-// int isDirectory( const std::string& full_path ) {
-// 	struct stat path_stat;
-// 	if (stat(full_path.c_str(), &path_stat) == 0) {
-// 		return (S_ISDIR(path_stat.st_mode));
-// 	}
-// 	return 0;
-// }
-
-// std::string getModTime( const std::string& path ) {
-// 	struct stat file_stat;
-// 	if (stat(path.c_str(), &file_stat) == 0) {
-// 		struct tm *tm = localtime(&file_stat.st_mtime);
-// 		char timebuf[80];
-// 		strftime(timebuf, sizeof(timebuf), "%d-%b-%Y %H:%M", tm);
-// 		return std::string(timebuf);
-// 	}
-// 	return "Unknown";
-// }
-
-// void Webserv::_generateDirectoryList( const std::string &dir_path, int client_fd ) {
-// 	std::ostringstream html;
-// 	std::string name, full_path, size, mod_time;
-// 	std::string& response = _clients_map[client_fd].response;
-// 	std::string& file_path = _clients_map[client_fd].request.path;
-
-// 	html << "<html><body><h1>Index of " << file_path << "</h1>\n<hr><pre><table>\n";
-
-// 	DIR *dir = opendir(dir_path.c_str());
-// 	if (dir == nullptr) {
-// 		_prepareResponseError(_clients_map[client_fd], 403);
-// 		logger.warning("Failed to open directory" + dir_path);
-// 		return;
-// 	}
-// 	struct dirent *entry;
-// 	while ((entry = readdir(dir)) != nullptr) {
-// 		name = entry->d_name;
-// 		full_path = dir_path + name;
-// 		if (name == ".") continue;
-// 		if (name == "..") {
-// 			name += "/";
-// 		} else if (entry->d_type == DT_DIR) {
-// 			name += "/";
-// 			size = "-";
-// 			mod_time = getModTime(full_path);
-// 		} else {
-// 			struct stat st;
-// 			stat(full_path.c_str(), &st);
-// 			size = std::to_string(st.st_size);
-// 			mod_time = getModTime(full_path);
-// 		}
-// 		html << "<tr><td style=\"width:70%\"><a href=\"" << name << "\">" << name << "</a></td>"
-// 			 << "<td style=\"width:20%\">" << mod_time << "</td>"
-// 			 << "<td align=\"right\">" << size << "</td></tr>\n";
-// 	}
-// 	closedir(dir);
-// 	html << "</table>\n</pre><hr></body>\n</html>\n";
-// 	response = "HTTP/1.1 200 OK\r\nContent-Type: text/html\r\nContent-Length: ";
-// 	response += std::to_string(html.str().size())+ "\r\n\r\n" + html.str();
-// }
-
-// int Webserv::_prepareResponse( int client_fd, const std::string& file_path, size_t status_code ) {
-// 	std::string& response = _clients_map[client_fd].response;
-// 	std::string& root_path = _clients_map[client_fd].location->root;
-// 	std::string full_path = root_path + file_path;
-// 	if (file_path == _clients_map[client_fd].location->path && !_clients_map[client_fd].location->index_page.empty()) {
-// 		std::string index_page = "/" + _clients_map[client_fd].location->index_page;
-// 		if (access((root_path + index_page).c_str(), F_OK) == 0) {
-// 			return _prepareResponse(client_fd, index_page, 200);
-// 		}
-// 	}
-// 	if (full_path.back() == '/' && isDirectory(full_path)) {
-// 		if (_clients_map[client_fd].location->autoindex == 1) {
-// 			_generateDirectoryList(full_path, client_fd);
-// 		} else {
-// 			_prepareResponseError(_clients_map[client_fd], 403);
-// 		}
-// 		return 1;
-// 	}
-// 	std::string extension = _getFileExtension(file_path);
-// 	if (extension == "py") {
-// 		logger.info("CGI file: " + full_path);
-// 		if (access(full_path.c_str(), X_OK) == 0) {
-// 		 	return _executeCgi(client_fd, full_path);
-// 		} else {
-// 			_prepareResponseError(_clients_map[client_fd], 404);
-// 			return 1;
-// 		}
-// 	}
-// 	std::ifstream file(full_path);
-// 	if (isDirectory(full_path) || !file.is_open()) {
-// 		logger.warning("Failed to open file: " + full_path);
-// 		_prepareResponseError(_clients_map[client_fd], 404);
-// 		return 1;
-// 	}
-// 	std::stringstream buffer;
-// 	buffer << file.rdbuf();
-// 	response = buffer.str();
-// 	response = _getHtmlHeader(response.size(), status_code, extension) + response;
-// 	return 1;
-// }
-
-// void Webserv::_prepareResponseError( ClientData& client_data, size_t status_code ) {
-// 	std::string filepath = "";
-// 	if (client_data.location != nullptr) {
-// 		filepath = _getErrorPagePath(client_data.location->error_pages, status_code);
-// 	}
-// 	if (!filepath.empty() && _saveResponsePage(client_data, filepath, status_code) == 0) {
-// 		return;
-// 	}
-// 	filepath = _getErrorPagePath(_error_pages, status_code);
-// 	if (filepath.empty()) {
-// 		filepath = _error_pages.at(0);
-// 	}
-// 	_saveResponsePage(client_data, filepath, status_code);
-// }
-
-// std::string Webserv::_getErrorPagePath(const std::unordered_map<int, std::string>& error_pages, size_t status_code) {
-// 	if (error_pages.find(status_code) == error_pages.end()) {
-// 		return "";
-// 	}
-// 	const std::string& filepath = error_pages.at(status_code);
-// 	if (access(filepath.c_str(), R_OK) == 0 && isDirectory(filepath) == 0) {
-// 		return filepath;
-// 	}
-// 	return "";
-// }
-
-// int Webserv::_saveResponsePage( ClientData& client_data, std::string& filepath, int status_code ) {
-// 	std::ifstream file(filepath);
-// 	if (!file.is_open()) {
-// 		return 1;
-// 	}
-// 	std::stringstream buffer;
-// 	buffer << file.rdbuf();
-// 	std::string& response = client_data.response;
-// 	response = buffer.str();
-// 	std::string extension = _getFileExtension(filepath);
-// 	response = _getHtmlHeader(response.size(), status_code, extension) + response;
-// 	return 0;
-// }
-
-// std::string Webserv::_getFileExtension( const std::string& filepath ) {
-// 	size_t pos = filepath.rfind('.');
-// 	if (pos == std::string::npos) {
-// 		return "";
-// 	}
-// 	return filepath.substr(pos + 1);
-// }
-
-// std::string Webserv::_getHtmlHeader( size_t content_length, size_t status_code, const std::string& extension ) {
-// 	std::string header = "HTTP/1.1 ";
-// 	if (_response_codes.find(status_code) != _response_codes.end()) {
-// 		header += _response_codes.at(status_code) + "\r\n";
-// 	} else {
-// 		header += std::to_string(status_code) + "\r\n";
-// 	}
-// 	std::string content_type;
-// 	if (_mime_types.find(extension) == _mime_types.end()) {
-// 		content_type = _mime_types.at("");
-// 	} else {
-// 		content_type = _mime_types.at(extension);
-// 	}
-// 	header += "Content-Type: " + content_type + "\r\n";
-// 	header += "Content-Length: " + std::to_string(content_length) + "\r\n\r\n";
-// 	return header;
-// }
-
 void Webserv::_getTargetServer(int client_fd, const std::string& host) {
 	ClientData& client_data = _clients_map[client_fd];
 	size_t delimiter = host.find(":");
