+++ Webserv.hpp
@@ -13,16 +13,13 @@
 #include <sys/socket.h>
 #include <sys/types.h>
 #include <sys/wait.h>
-#include <sys/stat.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
-#include <unistd.h>
 #include <list>
 #include <set>
 #include <unordered_map>
 #include <vector>
 #include <algorithm>
-#include <dirent.h>
 
 #include "Location.hpp"
 #include "Logger.hpp"
@@ -84,14 +81,7 @@ private:
 	void _stopServer( void );
 	void _mainLoop( void );
 	void _checkTimeouts( void );
-	// std::string _getHtmlHeader( size_t content_length, size_t status_code, const std::string& extension );
-	// int _prepareResponse( int client_fd, const std::string& file_path, size_t status_code = 200 );
-	// void _generateDirectoryList( const std::string &dir_path, int client_fd );
 	void _getTargetServer(int client_fd, const std::string& host);
-	// void _prepareResponseError( ClientData& client_data, size_t status_code );
-	// std::string _getErrorPagePath(const std::unordered_map<int, std::string>& error_pages, size_t status_code);
-	// int _saveResponsePage( ClientData& client_data, std::string& filepath, int status_code );
-	// std::string _getFileExtension( const std::string& filepath );
 
 	// WebservConfig.cpp
 	void _fakeConfigParser( void );
@@ -132,7 +122,6 @@ private:
 	int _setNonBlocking( int fd );
 	void _closeClientFd( int client_fd, const char* err_msg );
 	void _modifyEpollSocketOut( int client_fd );
-	// int _stringToInt( const std::string& str );
 
 public:
 	Webserv( const Webserv& ) = delete;
